Added the ability for the player guard to catch the the thief and the muse
