+++ TheThief-TheMuse2D/Assets/Characters/GuardPlayer/GuardPlayer.prefab
@@ -187,6 +187,36 @@ MonoBehaviour:
   m_RequiresDepthTexture: 0
   m_RequiresColorTexture: 0
   m_Version: 2
+--- !u!1 &8194887786931627048
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 2355826104180897891}
+  m_Layer: 0
+  m_Name: Guard View
+  m_TagString: Untagged
+  m_Icon: {fileID: 5721338939258241955, guid: 0000000000000000d000000000000000, type: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!4 &2355826104180897891
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 8194887786931627048}
+  m_LocalRotation: {x: -0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: -5.66, y: 2.8199997, z: 0}
+  m_LocalScale: {x: -1, y: 1, z: 1}
+  m_Children: []
+  m_Father: {fileID: 8523712483562564638}
+  m_RootOrder: 5
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &8523712482349273202
 GameObject:
   m_ObjectHideFlags: 0
@@ -328,6 +358,7 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 8523712483562564638}
+  - component: {fileID: 6042053346412395849}
   - component: {fileID: 8523712483562564612}
   - component: {fileID: 8523712483562564613}
   - component: {fileID: 8523712483562564632}
@@ -355,9 +386,27 @@ Transform:
   - {fileID: 8523712484070366785}
   - {fileID: 8523712482527708100}
   - {fileID: 5638522539950119427}
+  - {fileID: 2355826104180897891}
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &6042053346412395849
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 8523712483562564634}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: e6fd5e52008ea6948a0561677b70bedf, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  GuardView: {fileID: 2355826104180897891}
+  GuardViewRadius: 6
+  WhatIsPlayerLayer:
+    serializedVersion: 2
+    m_Bits: 512
 --- !u!114 &8523712483562564612
 MonoBehaviour:
   m_ObjectHideFlags: 0
