+++ TheThief-TheMuse2D/Assets/Characters/guard/Guard.prefab
@@ -59,7 +59,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 3318698057538446048}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: -5.19, y: 2.82, z: 0}
+  m_LocalPosition: {x: -5.72, y: 2.82, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 7838100107477081161}
@@ -104,6 +104,7 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 7838100107477081161}
+  - component: {fileID: 6118799320077103943}
   - component: {fileID: 7838100107477081162}
   - component: {fileID: 7838100107477081163}
   - component: {fileID: 7838100107477081160}
@@ -131,6 +132,23 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &6118799320077103943
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 7838100107477081166}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: e6fd5e52008ea6948a0561677b70bedf, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  GuardView: {fileID: 7649732942830349556}
+  GuardViewRadius: 6
+  WhatIsPlayerLayer:
+    serializedVersion: 2
+    m_Bits: 512
 --- !u!114 &7838100107477081162
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -143,15 +161,8 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 3c9284ad9b4e9a648bfd26194c459062, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  Player1Restart: {x: -40, y: -17}
-  Player2Restart: {x: -46, y: -17}
   speed: 8
   jumpForce: 5
-  GuardView: {fileID: 7649732942830349556}
-  GuardViewRadius: 5
-  WhatIsPlayerLayer:
-    serializedVersion: 2
-    m_Bits: 512
   groundCheck: {fileID: 7838100108912462894}
   wallCheck: {fileID: 7838100109329928885}
   checkRadius: 0.5
@@ -159,6 +170,7 @@ MonoBehaviour:
     serializedVersion: 2
     m_Bits: 256
   rb: {fileID: 0}
+  guardActions: {fileID: 0}
 --- !u!50 &7838100107477081163
 Rigidbody2D:
   serializedVersion: 4
