+++ TheThief-TheMuse2D/Assets/Scenes/TestScene.unity
@@ -1283,7 +1283,7 @@ GameObject:
   - component: {fileID: 682638224}
   m_Layer: 0
   m_Name: PlayerInputManager
-  m_TagString: Untagged
+  m_TagString: PlayerInputManager
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
@@ -1365,6 +1365,8 @@ MonoBehaviour:
     type: 3}
   PlayerPrefabGuard: {fileID: 8523712483562564634, guid: 1e7db8b822d366a419b8e405ef55d169,
     type: 3}
+  Player1Restart: {x: -45, y: -17}
+  Player2Restart: {x: -40, y: -17}
 --- !u!1 &689212407
 GameObject:
   m_ObjectHideFlags: 0
