+++ TheThief-TheMuse2D/Assets/Scripts/GaurdController.cs
@@ -5,19 +5,11 @@ using UnityEngine;
 
 public class GaurdController : MonoBehaviour
 {
-    public Vector2 Player1Restart;
-    public Vector2 Player2Restart;
-
     public float speed = 10;
     public float jumpForce = 5;
     private float moveInput = -1;
 
     private bool facingRight = true;
-
-    private bool isPlayerInGuardView;
-    public Transform GuardView;
-    public float GuardViewRadius = 10f;
-    public LayerMask WhatIsPlayerLayer;
 
     private bool isGrounded;
     public Transform groundCheck;
@@ -32,25 +24,15 @@ public class GaurdController : MonoBehaviour
 
     public Rigidbody2D rb;
     // Start is called before the first frame update
+    public GuardActions guardActions;
     void Start()
     {
         rb = GetComponent<Rigidbody2D>();
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        
+        guardActions = GetComponent<GuardActions>();
     }
 
     private void FixedUpdate()
     {
-        isPlayerInGuardView = Physics2D.OverlapCircle(GuardView.position, GuardViewRadius, WhatIsPlayerLayer);
-        if(isPlayerInGuardView)
-        {
-            PlayerInGaurdView();
-        }
-
         isGrounded = Physics2D.OverlapCircle(groundCheck.position, checkRadius, whatIsGround);
         isWallChecked = Physics2D.OverlapCircle(wallCheck.position, checkRadius, whatIsGround);
         if (!isGrounded || isWallChecked)
@@ -58,21 +40,13 @@ public class GaurdController : MonoBehaviour
             Flip();
             moveInput *= -1;
         }
-
         rb.velocity = new Vector2(moveInput * speed, rb.velocity.y);
-
-
     }
 
-    private void PlayerInGaurdView()
-    {
-        Debug.Log("Caughtttttttttttttttttttttt Start over");
-        GameObject[] allPlayers = GameObject.FindGameObjectsWithTag("Player");
-
-        allPlayers[0].transform.position = new Vector3(Player1Restart.x, Player1Restart.y);
-        allPlayers[1].transform.position = new Vector3(Player2Restart.x, Player2Restart.y);
-
-    }
+    //private void IsPlayerInGaurdView()
+    //{
+    //    guardActions.IsPlayerInGaurdView(GuardView.position, GuardViewRadius, WhatIsPlayerLayer);
+    //}
 
     void Flip()
     {
@@ -84,13 +58,6 @@ public class GaurdController : MonoBehaviour
 
     }
 
-    private void OnDrawGizmos()
-    {
-//        Gizmos.DrawSphere(GuardView.position, GuardViewRadius);
-
-        Gizmos.DrawWireSphere(GuardView.position, GuardViewRadius);
-
-    }
 
 
 }
