+++ TheThief-TheMuse2D/ProjectSettings/TagManager.asset
@@ -3,7 +3,8 @@
 --- !u!78 &1
 TagManager:
   serializedVersion: 2
-  tags: []
+  tags:
+  - PlayerInputManager
   layers:
   - Default
   - TransparentFX
